#26 Room과 ChatModule 분리
Room과 ChatModule 분리
