+++ bubblit/assets/js/Component/Room.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import { Divider, Input, Button } from 'semantic-ui-react'
+import ChatModule from './ChatModule'
 
 export default class Chat extends Component {
     constructor(props) {
@@ -9,6 +10,7 @@ export default class Chat extends Component {
             chatData: '',
         }
     }
+
     chatRenderer() {
         var i = 0;
         var content = [];
@@ -22,31 +24,15 @@ export default class Chat extends Component {
         }
         return content
     }
+
     render() {
         return (
             <div>
                 <h2>Room '{this.props.mode}'</h2>
-                {this.chatRenderer()}
-                <Input
-                    name='ChatInput'
-                    onChange={function (e) {
-                        this.setState({ chatData: e.target.value })
-                    }.bind(this)}
-                    action={{
-                        icon: 'arrow up',
-                        onClick: function (e, data) {
-                            this.props.onClick(this.state.chatData, this.state.userID, Date());
-                            //this.setState({ chatData: '' });
-                            //var input = document.getElementsByName('chatInput');
-                            //input[0].value = ''; //e, data를 활용하여 초기화하는 방법이 있을것같음. 추후 수정
-                        }.bind(this)
-                    }}
-                ></Input>
-                <p></p>
-                <Button onClick={function (e, data) {
-                    this.props.exitRoom();
-                }.bind(this)}
-                >Exit Room</Button>
+                <ChatModule
+                    chatDB={this.props.chatDB}
+                    onClick={this.props.onClick}
+                ></ChatModule>
             </div >
         )
     }
