+++ bubblit/assets/js/app.js
@@ -25,12 +25,9 @@ import Board from './Component/Board'
 
 // This code starts up the React app when it runs in a browser. It sets up the routing
 // configuration and injects the app into a DOM element.
-ReactDOM.render(<ChatTestModule />, document.getElementById('react-juno'))
+//ReactDOM.render(<ChatTestModule />, document.getElementById('react-juno'))
 
-/*
-프로토타입 제작을 위한 주석 처리
-*/
-// ReactDOM.render(
-//     <Provider store={store}>
-//         <Board />
-//     </Provider>, document.getElementById('react-app'))
+ReactDOM.render(
+    <Provider store={store}>
+        <Board />
+    </Provider>, document.getElementById('react-app'))
